Pools can mint directly. Stakers can still claim rewards from deactivated pools.
