+++ clarity/tests/arkadiko-stake-registry-v1-1_test.ts
@@ -29,7 +29,6 @@ Clarinet.test({
       types.principal('STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-pool-diko-v1-1')
     ], wallet_1.address);
     call.result.expectTuple()['name'].expectAscii('DIKO');
-    call.result.expectTuple()['active'].expectBool(true);
   }
 });
 
@@ -922,4 +921,291 @@ Clarinet.test({
     call = chain.callReadOnlyFn("arkadiko-stake-pool-diko-xusd-v1-1", "get-total-staked", [], deployer.address);
     call.result.expectUint(0);
   }
-  });
+});
+
+Clarinet.test({
+  name: "dao: replace pool through new registry",
+  async fn(chain: Chain, accounts: Map<string, Account>) {
+    let deployer = accounts.get("deployer")!;
+    let wallet_1 = accounts.get("wallet_1")!;
+
+    // Stake funds
+    let block = chain.mineBlock([
+      Tx.contractCall("arkadiko-stake-registry-v1-1", "stake", [
+        types.principal('STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-registry-v1-1'),
+        types.principal('STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-pool-diko-v1-1'),
+        types.principal('STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-token'),
+        types.uint(100000000)
+      ], deployer.address)
+    ]);
+    block.receipts[0].result.expectOk().expectUint(100000000); // 10 with 6 decimals
+
+    // Create proposal
+    block = chain.mineBlock([
+    Tx.contractCall("arkadiko-governance-v1-1", "propose", [
+      types.uint(10),
+      types.utf8("change-reward-distribution"),
+      types.list([
+        types.tuple({
+          'name': types.ascii("stake-registry"),
+          'address': types.principal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7"),
+          'qualified-name': types.principal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-registry-tv1-1"),
+          'can-mint': types.bool(true),
+          'can-burn': types.bool(true)
+        }),
+        types.tuple({
+          'name': types.ascii("stake-pool-diko-2"),
+          'address': types.principal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7"),
+          'qualified-name': types.principal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-pool-diko-tv1-1"),
+          'can-mint': types.bool(true),
+          'can-burn': types.bool(true)
+        })
+      ])
+    ], wallet_1.address)
+    ]);
+    block.receipts[0].result.expectOk().expectBool(true);
+
+    let call:any = chain.callReadOnlyFn("arkadiko-governance-v1-1", "get-proposal-by-id", [types.uint(1)], wallet_1.address);
+    call.result.expectTuple()["is-open"].expectBool(true);
+    call.result.expectTuple()["start-block-height"].expectUint(10);
+    
+    // Advance
+    chain.mineEmptyBlock(10);
+
+    // Vote for wallet_1
+    block = chain.mineBlock([
+    Tx.contractCall("arkadiko-governance-v1-1", "vote-for", [
+        types.principal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-token"),
+        types.uint(1),
+        types.uint(10000000)
+    ], deployer.address)
+    ]);
+    block.receipts[0].result.expectOk().expectUint(3200);
+
+    // Advance
+    chain.mineEmptyBlock(1500);
+
+    // End proposal
+    block = chain.mineBlock([
+    Tx.contractCall("arkadiko-governance-v1-1", "end-proposal", [
+        types.uint(1)
+    ], deployer.address)
+    ]);
+    block.receipts[0].result.expectOk().expectUint(3200);
+
+    // Check if DAO updated
+    call = chain.callReadOnlyFn("arkadiko-dao", "get-contract-address-by-name", [types.ascii("stake-registry")], deployer.address);
+    call.result.expectSome().expectPrincipal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7");
+    call = chain.callReadOnlyFn("arkadiko-dao", "get-qualified-name-by-name", [types.ascii("stake-registry")], deployer.address);
+    call.result.expectSome().expectPrincipal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-registry-tv1-1");
+
+    // Check mint and burn authorisation
+    call = chain.callReadOnlyFn("arkadiko-dao", "get-contract-can-mint-by-qualified-name", [types.principal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-registry-v1-1")], deployer.address);
+    call.result.expectBool(false)
+    call = chain.callReadOnlyFn("arkadiko-dao", "get-contract-can-burn-by-qualified-name", [types.principal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-registry-v1-1")], deployer.address);
+    call.result.expectBool(false)
+      
+    call = chain.callReadOnlyFn("arkadiko-dao", "get-contract-can-mint-by-qualified-name", [types.principal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-registry-tv1-1")], deployer.address);
+    call.result.expectBool(true)
+    call = chain.callReadOnlyFn("arkadiko-dao", "get-contract-can-burn-by-qualified-name", [types.principal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-registry-tv1-1")], deployer.address);
+    call.result.expectBool(true)
+
+    call = chain.callReadOnlyFn("arkadiko-dao", "get-contract-can-mint-by-qualified-name", [types.principal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-pool-diko-v1-1")], deployer.address);
+    call.result.expectBool(true)
+    call = chain.callReadOnlyFn("arkadiko-dao", "get-contract-can-burn-by-qualified-name", [types.principal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-pool-diko-v1-1")], deployer.address);
+    call.result.expectBool(true)
+      
+    call = chain.callReadOnlyFn("arkadiko-dao", "get-contract-can-mint-by-qualified-name", [types.principal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-pool-diko-tv1-1")], deployer.address);
+    call.result.expectBool(true)
+    call = chain.callReadOnlyFn("arkadiko-dao", "get-contract-can-burn-by-qualified-name", [types.principal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-pool-diko-tv1-1")], deployer.address);
+    call.result.expectBool(true)
+
+    // Pending rewards on old pool
+    call = chain.callReadOnlyFn("arkadiko-stake-pool-diko-v1-1", "get-pending-rewards", [
+      types.principal('STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-registry-tv1-1'),
+      types.principal(deployer.address)
+    ], deployer.address);
+    call.result.expectOk().expectUint(1199400)
+
+    // Advance
+    chain.mineEmptyBlock(1500);
+
+    // Pending rewards on old pool stay the same
+    call = chain.callReadOnlyFn("arkadiko-stake-pool-diko-v1-1", "get-pending-rewards", [
+      types.principal('STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-registry-tv1-1'),
+      types.principal(deployer.address)
+    ], deployer.address);
+    call.result.expectOk().expectUint(1199400)
+
+    // Stake funds fails as pool is not active anymore
+    block = chain.mineBlock([
+      Tx.contractCall("arkadiko-stake-registry-tv1-1", "stake", [
+        types.principal('STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-registry-tv1-1'),
+        types.principal('STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-pool-diko-v1-1'),
+        types.principal('STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-token'),
+        types.uint(100000000)
+      ], deployer.address)
+    ]);
+    block.receipts[0].result.expectErr().expectUint(19003);
+
+    // Wrong registry as parameter
+    block = chain.mineBlock([
+      Tx.contractCall("arkadiko-stake-registry-tv1-1", "stake", [
+        types.principal('STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-registry-v1-1'),
+        types.principal('STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-pool-diko-v1-1'),
+        types.principal('STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-token'),
+        types.uint(100000000)
+      ], deployer.address)
+    ]);
+    block.receipts[0].result.expectErr().expectUint(19004);
+
+    // Unstake funds with old registry should fail
+    block = chain.mineBlock([
+      Tx.contractCall("arkadiko-stake-registry-tv1-1", "unstake", [
+        types.principal('STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-registry-v1-1'),
+        types.principal('STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-pool-diko-v1-1'),
+        types.principal('STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-token'),
+        types.uint(100000000)
+      ], deployer.address)
+    ]);
+    block.receipts[0].result.expectErr().expectUint(19004)
+
+    // Unstake funds with new registry succeeds
+    block = chain.mineBlock([
+      Tx.contractCall("arkadiko-stake-registry-tv1-1", "unstake", [
+        types.principal('STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-registry-tv1-1'),
+        types.principal('STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-pool-diko-v1-1'),
+        types.principal('STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-token'),
+        types.uint(100000000)
+      ], deployer.address)
+    ]);
+    block.receipts[0].result.expectOk().expectUint(100000000)
+
+  }
+});
+
+Clarinet.test({
+  name: "dao: shut down pool because of critical bug",
+  async fn(chain: Chain, accounts: Map<string, Account>) {
+    let deployer = accounts.get("deployer")!;
+    let wallet_1 = accounts.get("wallet_1")!;
+
+    // Stake funds
+    let block = chain.mineBlock([
+      Tx.contractCall("arkadiko-stake-registry-v1-1", "stake", [
+          types.principal('STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-registry-v1-1'),
+          types.principal('STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-pool-diko-v1-1'),
+          types.principal('STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-token'),
+          types.uint(100000000)
+      ], deployer.address)
+    ]);
+    block.receipts[0].result.expectOk().expectUint(100000000); // 10 with 6 decimals
+
+    // Create proposal
+    block = chain.mineBlock([
+    Tx.contractCall("arkadiko-governance-v1-1", "propose", [
+      types.uint(10),
+      types.utf8("change-reward-distribution"),
+      types.list([
+        types.tuple({
+          'name': types.ascii("stake-registry"),
+          'address': types.principal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7"),
+          'qualified-name': types.principal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-registry-tv1-1"),
+          'can-mint': types.bool(true),
+          'can-burn': types.bool(true)
+        }),
+        types.tuple({
+          'name': types.ascii("stake-pool-diko"),
+          'address': types.principal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7"),
+          'qualified-name': types.principal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-pool-diko-v1-1"),
+          'can-mint': types.bool(false),
+          'can-burn': types.bool(true)
+        }),
+        types.tuple({
+          'name': types.ascii("stake-pool-diko-2"),
+          'address': types.principal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7"),
+          'qualified-name': types.principal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-pool-diko-tv1-1"),
+          'can-mint': types.bool(true),
+          'can-burn': types.bool(true)
+        })
+      ])
+    ], wallet_1.address)
+    ]);
+    block.receipts[0].result.expectOk().expectBool(true);
+
+    let call:any = chain.callReadOnlyFn("arkadiko-governance-v1-1", "get-proposal-by-id", [types.uint(1)], wallet_1.address);
+    call.result.expectTuple()["is-open"].expectBool(true);
+    call.result.expectTuple()["start-block-height"].expectUint(10);
+    
+    // Advance
+    chain.mineEmptyBlock(10);
+
+    // Vote for wallet_1
+    block = chain.mineBlock([
+    Tx.contractCall("arkadiko-governance-v1-1", "vote-for", [
+        types.principal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-token"),
+        types.uint(1),
+        types.uint(10000000)
+    ], deployer.address)
+    ]);
+    block.receipts[0].result.expectOk().expectUint(3200);
+
+    // Advance
+    chain.mineEmptyBlock(1500);
+
+    // End proposal
+    block = chain.mineBlock([
+    Tx.contractCall("arkadiko-governance-v1-1", "end-proposal", [
+        types.uint(1)
+    ], deployer.address)
+    ]);
+    block.receipts[0].result.expectOk().expectUint(3200);
+
+    // Check if DAO updated
+    call = chain.callReadOnlyFn("arkadiko-dao", "get-contract-address-by-name", [types.ascii("stake-registry")], deployer.address);
+    call.result.expectSome().expectPrincipal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7");
+    call = chain.callReadOnlyFn("arkadiko-dao", "get-qualified-name-by-name", [types.ascii("stake-registry")], deployer.address);
+    call.result.expectSome().expectPrincipal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-registry-tv1-1");
+
+    // Check mint and burn authorisation
+    call = chain.callReadOnlyFn("arkadiko-dao", "get-contract-can-mint-by-qualified-name", [types.principal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-registry-v1-1")], deployer.address);
+    call.result.expectBool(false)
+    call = chain.callReadOnlyFn("arkadiko-dao", "get-contract-can-burn-by-qualified-name", [types.principal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-registry-v1-1")], deployer.address);
+    call.result.expectBool(false)
+      
+    call = chain.callReadOnlyFn("arkadiko-dao", "get-contract-can-mint-by-qualified-name", [types.principal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-registry-tv1-1")], deployer.address);
+    call.result.expectBool(true)
+    call = chain.callReadOnlyFn("arkadiko-dao", "get-contract-can-burn-by-qualified-name", [types.principal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-registry-tv1-1")], deployer.address);
+    call.result.expectBool(true)
+
+    call = chain.callReadOnlyFn("arkadiko-dao", "get-contract-can-mint-by-qualified-name", [types.principal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-pool-diko-v1-1")], deployer.address);
+    call.result.expectBool(false)
+    call = chain.callReadOnlyFn("arkadiko-dao", "get-contract-can-burn-by-qualified-name", [types.principal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-pool-diko-v1-1")], deployer.address);
+    call.result.expectBool(true)
+      
+    call = chain.callReadOnlyFn("arkadiko-dao", "get-contract-can-mint-by-qualified-name", [types.principal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-pool-diko-tv1-1")], deployer.address);
+    call.result.expectBool(true)
+    call = chain.callReadOnlyFn("arkadiko-dao", "get-contract-can-burn-by-qualified-name", [types.principal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-pool-diko-tv1-1")], deployer.address);
+    call.result.expectBool(true)
+
+    // Unstake funds should fail as pool is not able to mint
+    block = chain.mineBlock([
+      Tx.contractCall("arkadiko-stake-registry-tv1-1", "unstake", [
+        types.principal('STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-registry-tv1-1'),
+        types.principal('STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-pool-diko-v1-1'),
+        types.principal('STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-token'),
+        types.uint(100000000)
+      ], deployer.address)
+    ]);
+    block.receipts[0].result.expectErr().expectUint(100401)
+
+    // Can use emergency withdraw method
+    block = chain.mineBlock([
+      Tx.contractCall("arkadiko-stake-pool-diko-v1-1", "emergency-withdraw", [
+        types.principal('STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-stake-registry-tv1-1')
+      ], deployer.address)
+    ]);
+    block.receipts[0].result.expectOk().expectUint(0)
+
+  }
+});
